# lab2example/src/main/java/info/stepanoff/trsis/samples/rest/model/LessonDataWithoutID.java
package info.stepanoff.trsis.samples.rest.model;

import lombok.Getter;

@Getter
public class LessonDataWithoutID {
    private String discipline;
    private String lessonType;
    private String audience;
    private String address;
    private String start;

}